requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
